My original code with the training added in
I know the code is a little messy and I wanted to make the training into a function at least but for some reason, the error tripled when I did that. I have done three training sessions and depending on random variables with an eta of 0.1 -- anything lower just increases the error -- I can get the big E error down to 0.03. The first training session uses the weights from the class, the second training uses random weights, and the third training uses a random combination of the final weights from the first two sessions.